Added script to calculate the token cost of sending to ChatGPT.
I don't have an answer key for the test data set in this repository.
Without the answer key, models I create can't be evaluated with this
dataset. I could manually read the texts and label them, but there are
11,000 files.

I see this as a good opportunity to learn and use the chat-gpt api. Is
this within the terms and conditions? ...I asked GPT and got no
complaints!

To use the API, a model must be specified. GPT-4 is in limited beta, so
I'll use gpt-3.5-turbo. This model may change, so it will be better to
use a tagged version, such as gpt-3.5-turbo-0301.

API calls incur a cost. The cost is counted in tokens sent and received.
A single call can have a maximum of 4096 tokens. My account is granted
$18 until June 1 2023. It doesn't look like anything has been spent so
far!

This presents a potential problem: could the movie reviews I want to
send it be too long?

The script in this commit checks the cost of each review. This is
written to a csv file for me to process later. A cursory look at the csv
shows that none of the files ever breach 2000 tokens, so the API calls
will be safe!

The script was made with the help of Chat GPT. It's been a while since
I've written python code. I've forgotten how to do many common
operations, but I know what can be done. GPT has been a great help in
this regard. I learned a lot, and even corrected a provided example. The
overall experience was highly engaging. I imagine this is what peer
programming is like, but interacting with GPT has the advantage of no
social pressue.

I stay organized with the GPT conversations:
  [vim help]
  [nm-cli help]
  [python help]
This helps maintain the context of the conversation to a certain topic.
Not sure if it makes a difference on the GPT side, but it can't hurt...
plus it helps me reference past interactions.